Remove Dashboard component and its associated logic
- Deleted the Dashboard.tsx file, which contained the main task management interface, including state management, task filtering, and import/export functionality.
- This change is part of a broader effort to modularize the application and improve code organization.
